credit card updated

+++ src/pages/SelectedClasses/SelectedClassesTable/SelectedClassesTable.jsx
@@ -3,10 +3,38 @@ import SelectedClassesTableHead from "./SelectedClassesTableHead";
 import { BsFillCreditCardFill, BsFillTrash3Fill } from "react-icons/bs";
 import { deleteAllClass, deleteClass } from "../../../api/bookApi";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
+import { toast } from "react-toastify";
 
 const SelectedClassesTable = ({ userBookings, userDetails }) => {
   const handleClearList = () => {
-    deleteAllClass(userDetails._id);
+    Swal.fire({
+      title: "Are you sure that you want to clear your booking list?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      color: "white",
+      iconColor: "rgb(234 179 8)",
+      showCancelButton: true,
+      confirmButtonColor: "rgb(234 179 8)",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, clear list!",
+      background: "#201e1e",
+      backdrop: "#00000",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Booking List has been Cleared!",
+          icon: "success",
+          color: "white",
+          iconColor: "lightgreen",
+          confirmButtonColor: "lightgreen",
+          confirmButtonText: "OK",
+          background: "#201e1e",
+          backdrop: "#00000",
+        });
+        deleteAllClass(userDetails._id);
+      }
+    });
   };
 
   if (userBookings?.length === 0) {
@@ -18,6 +46,31 @@ const SelectedClassesTable = ({ userBookings, userDetails }) => {
       </div>
     );
   }
+
+  const updateProfile = () => {
+    if (
+      typeof userDetails.address === "undefined" ||
+      typeof userDetails.contactNo === "undefined" ||
+      userDetails.gender === "undefined"
+    ) {
+      toast.warning(
+        "To purchase classes, you have to update your profile first!",
+        {
+          position: "top-center",
+          autoClose: 1500,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: "dark",
+        }
+      );
+      setTimeout(function () {
+        window.location.replace("/dashboard/profile");
+      }, 2400);
+    }
+  };
 
   return (
     <div className="overflow-x-auto">
@@ -45,9 +98,7 @@ const SelectedClassesTable = ({ userBookings, userDetails }) => {
 
             return (
               <tr className="" key={classItem._id}>
-                <td>
-                  {index + 1}
-                </td>
+                <td>{index + 1}</td>
                 <td className="flex justify-center">
                   <img
                     className="w-16 rounded-xl h-8"
@@ -66,8 +117,14 @@ const SelectedClassesTable = ({ userBookings, userDetails }) => {
                 </td>
                 <td>$ {classItem.classFee}</td>
                 <td>
-                  <Link 
-                    to={`/dashboard/payment/${classItem.studentId}/${classItem._id}`}
+                  <Link
+                    onClick={updateProfile}
+                    to={
+                      userDetails.address &&
+                      userDetails.contactNo &&
+                      userDetails.gender &&
+                      `/dashboard/payment/${classItem.studentId}/${classItem._id}`
+                    }
                     className="btn text-white btn-xs text-sx border-0 rounded-lg hover:bg-stone-700 bg-stone-800"
                   >
                     <BsFillCreditCardFill /> <span>Pay</span>
